Block custom role from deletion if there are dependent policies

See merge request https://gitlab.com/gitlab-org/gitlab/-/merge_requests/176340

Merged-by: Gregory Havenga <11164960-ghavenga@users.noreply.gitlab.com>
Approved-by: Dominic Bauer <dbauer@gitlab.com>
Approved-by: Alper Akgun <aakgun@gitlab.com>
Approved-by: Gregory Havenga <11164960-ghavenga@users.noreply.gitlab.com>
Approved-by: mo khan <mo@mokhan.ca>

+++ ee/app/services/member_roles/delete_service.rb
@@ -7,6 +7,13 @@ module MemberRoles
 
       return authorized_error unless allowed?
 
+      if member_role.dependent_security_policies.exists?
+        return ::ServiceResponse.error(
+          message: 'Custom role linked with a security policy.',
+          payload: { member_role: member_role }
+        )
+      end
+
       if member_role.destroy
         log_audit_event(member_role, action: :deleted)
